saliency-service/requirements.txt
# CHECKION saliency service – HTTP API only; MDS-ViTNet deps live in mds_vitnet/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
numpy
pillow
torch>=1.13.1
torchvision>=0.14.1
